Use message time rather than server time
Previously, the date for "/mensa" requests was calculated by adding the date offset
to the current date at the server.
This lead to problems when responding to messages from users from a timezone
differing from the one of the server - people would get the menu for the last day at their
current position, because the date had not changed for the server yet.

Now, the date of the message is used.
Unfortunately, the telegram bot API does not (yet?) carry information about the user's timezone,
so all date times are in Unix Time, GMT.
However, this is still an improvement since the response from the bot should now return equal
responses, independently from where the message is sent from.